s/.ethers// s/_ethers//
