+++ unlock-js/src/v01/purchaseKey.js
@@ -1,10 +1,7 @@
-/* eslint no-console: 0 */
-
-import * as UnlockV01 from 'unlock-abi-0-1'
-import Web3Utils from '../utils'
+import Web3Utils from '../utils.ethers'
 import { GAS_AMOUNTS } from '../constants'
-import Errors from '../errors'
 import TransactionTypes from '../transactionTypes'
+import Errors from '../errors'
 
 /**
  * Purchase a key to a lock by account.
@@ -15,33 +12,23 @@ import TransactionTypes from '../transactionTypes'
  * @param {PropTypes.address} lock
  * @param {PropTypes.address} owner
  * @param {string} keyPrice
+ * @param {string} data
  * @param {string} account
- * @param {string} data // This field was previously available but has now been deprecated
  */
-export default function(lock, owner, keyPrice, account, data = '') {
-  if (data !== '') {
-    console.warn('The data field on purchaseFor has been deprecated.')
+export default async function(lockAddress, owner, keyPrice) {
+  const lockContract = await this.getLockContract(lockAddress)
+  let transactionPromise
+  try {
+    transactionPromise = lockContract['purchaseFor(address)'](owner, {
+      gasLimit: GAS_AMOUNTS.purchaseFor, // overrides default value for transaction gas price
+      value: Web3Utils.toWei(keyPrice, 'ether'), // overrides default value
+    })
+    const ret = await this._handleMethodCall(
+      transactionPromise,
+      TransactionTypes.KEY_PURCHASE
+    )
+    return ret
+  } catch (error) {
+    this.emit('error', new Error(Errors.FAILED_TO_PURCHASE_KEY))
   }
-  const lockContract = new this.web3.eth.Contract(
-    UnlockV01.PublicLock.abi,
-    lock
-  )
-  const abi = lockContract.methods.purchaseFor(owner).encodeABI()
-
-  return this._sendTransaction(
-    {
-      to: lock,
-      from: account,
-      data: abi,
-      gas: GAS_AMOUNTS.purchaseFor,
-      value: Web3Utils.toWei(keyPrice, 'ether'),
-      contract: UnlockV01.PublicLock,
-    },
-    TransactionTypes.KEY_PURCHASE,
-    error => {
-      if (error) {
-        return this.emit('error', new Error(Errors.FAILED_TO_PURCHASE_KEY))
-      }
-    }
-  )
 }
